requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
pymysql
python-dotenv
Flask-Migrate
gunicorn
cryptography


requests
